requirements.txt
pandas==1.5.3
scipy==1.10.1
biopython==1.81
numpy==1.24.2
requests==2.28.2
watchdog==2.3.1
streamlit==1.19.0
py3Dmol==2.0.0.post2
stmol==0.0.9